[imp][compiler] Pipeline compilation done per-file
Instead of building one large code file, then trying to parse to grab code out, we do the compilation as type-per-file.

This avoids the parsing of completed code completely, reduces the number of times GeneratedTypes is iterated, and improves stack traces etc. as the source is tied to the class that generates an exception