[Apm] Pull span management out of individual providers' middleware
APM packages no longer require middleware that is in essence a copy-paste between all of them. Instead the "hosts" of the APIs are expected to create / modify an IApmSpan and then place it in to the ApiOperationContext.

Individual modules can therefore add extra metadata (i.e. authorisation middleware), in addition to a single exception handling frame to push to the current span.

HTTP and Task hosts updated to set IApmSpan correctly.