[Compilation] Overhaul compilation to remove "PreCompilation" and introduce store-to-disk compilation
Following a similar pattern to Marten we will now store pipelines to disk during development, which will allow the types to be directly compiled-in as if they were hand-written, avoiding the pre-compilation dance on build and any dynamic assembly loading.

This also has the advantage that locally we will do as little compilation as possible by only re-compiling when strictly necessary in test and development runs, using existing compiled types in the assembly where possible.