[Http] Move the registration and loading of ApiOperationLink to HttpHost
This is the beginning of a refactor to how links / routes are handled. The first stage is moving the management of ApiOperationLinks to
Blueprint.Http completely, making the "base" ApiDataModel more host agnostic by removing the URL/HTTP focused links.

Work will continue to make declaring routes / links more consistent and easier to work with, especially in microservices with
operations that do not need a "nice" REST(ish) URL design and therefore could have "automatic" URL generation.