[Logging] Enforce use of explicit EventIds for more deterministic code generation
Using automatic event ids that are auto incremented means that if a new pipeline is created that is generated before the end of the list of pipelines, all subsequent pipelines would have different event ids and therefore require complete recompilation.

By enforcing explicit EventIds in logging calls we can make sure the resulting code is deterministic with regards to LogFrames and therefore avoid the problem.