[Auth] Always try to load authentication context, even when anonymous access allowed
There exist scenarios where even though an operation may be marked as AllowAnonymous it could have been called by an authenticated user and the command may wish to know.

In addition we will remove the default IClaimsIdentityProvider as we handle no registration in the AuthenticationMiddlewareBuilder to elide the check to the provider. This is typically the case as most use-cases do not have a specific loader but are instead set up in the context before running (as HTTP does for example)