[Handlers] Remove scanning assemblies for IApiOperationHandler and optimise all scan #breaking
Scanning for handlers was inefficient because each scanner was only given visibility to an operation at a time, meaning any global one-off optimisations were not possible.

This breaking change to the interface means that each scanner has the opportunity to perform these optimisations, which the ApiOperationHandlerExecutorBuilderScanner takes advantage of by scanning registered IoC services only once for handlers that have been registered, instead of having to do this search for every single operation.

In addition, the decision has been taken to remove manual scanning of assemblies for IApiOperationHandler implementations, instead delegating all to the IoC. This means that consumers may have to update their IoC registrations to do the scanning, instead of allowing Blueprint to do it implictly.

These changes have drastically reduced allocations and runtime when scanning for operations. A benchmark with 26 operations and handlers, with 500 services in the IoC container has reduced allocation by 70% and runtime by ~80%:

|    Method |       Mean |    Error |   StdDev | Ratio |    Gen0 |   Gen1 | Allocated | Alloc Ratio |
|---------- |-----------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
|      Base | 1,051.8 us | 12.44 us | 10.39 us |  1.00 | 27.3438 | 5.8594 | 116.44 KB |        1.00 |
| Optimised |   228.2 us |  1.33 us |  1.11 us |  0.22 |  7.0801 | 1.4648 |   35.2 KB |        0.30 |